rename FooterButtons in FooterWithButtons
