+++ ts/screens/wallet/AddCardScreen.tsx
@@ -21,7 +21,7 @@ import { NavigationScreenProp, NavigationState } from "react-navigation";
 import { LabelledItem } from "../../components/LabelledItem";
 import { WalletStyles } from "../../components/styles/wallet";
 import AppHeader from "../../components/ui/AppHeader";
-import FooterButtons from "../../components/ui/FooterButtons";
+import FooterWithButtons from "../../components/ui/FooterWithButtons";
 import IconFont from "../../components/ui/IconFont";
 import { cardIcons } from "../../components/wallet/card/Logo";
 import I18n from "../../i18n";
@@ -226,7 +226,7 @@ export class AddCardScreen extends React.Component<Props, State> {
           </Content>
         </ScrollView>
 
-        <FooterButtons
+        <FooterWithButtons
           leftButton={primaryButtonProps}
           rightButton={secondaryButtonProps}
           inlineHalf={true}
