+++ ts/theme/components/FooterWithButtons.ts
