+++ ts/theme/index.ts
@@ -7,7 +7,7 @@ import merge from "lodash/merge";
 import getTheme from "native-base/src/theme/components";
 import buttonTheme from "./components/Button";
 import contentTheme from "./components/Content";
-import footerButtonsTheme from "./components/FooterButtons";
+import footerWithButtonsTheme from "./components/FooterWithButtons";
 import h1Theme from "./components/H1";
 import h3Theme from "./components/H3";
 import h4Theme from "./components/H4";
@@ -93,8 +93,8 @@ const theme = (): Theme => {
     "UIComponent.ScreenHeader": {
       ...screenHeaderTheme()
     },
-    "UIComponent.FooterButtons": {
-      ...footerButtonsTheme()
+    "UIComponent.FooterWithButtons": {
+      ...footerWithButtonsTheme()
     },
     "NativeBase.ListItem": {
       ...listItemTheme()
